# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Supercheck API",
    "version": "1.0.0",
    "description": "API for test automation, monitoring, and CI/CD integration. Supercheck provides Playwright browser testing, k6 performance testing, uptime monitoring, and alerting — all managed via API or CLI.",
    "contact": {
      "name": "Supercheck",
      "url": "https://supercheck.io"
    },
    "license": {
      "name": "AGPL-3.0",
      "url": "https://github.com/supercheck-io/supercheck/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "https://app.supercheck.io",
      "description": "Production (Cloud)"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/monitors/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "getMonitor",
        "summary": "Get monitor details",
        "tags": [
          "Monitors"
        ],
        "responses": {
          "200": {
            "description": "Monitor details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Monitor"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "operationId": "updateMonitor",
        "summary": "Update a monitor",
        "tags": [
          "Monitors"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "url": {
                    "type": "string",
                    "format": "uri"
                  },
                  "interval": {
                    "type": "integer"
                  },
                  "timeout": {
                    "type": "integer"
                  },
                  "method": {
                    "type": "string"
                  },
                  "active": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Monitor updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Monitor"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "operationId": "deleteMonitor",
        "summary": "Delete a monitor",
        "tags": [
          "Monitors"
        ],
        "responses": {
          "200": {
            "description": "Monitor deleted"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "CLI token (`sck_live_*`) for full API access, or trigger key (`sck_trigger_*`) for job trigger endpoint only."
      }
    },
...
    "responses": {
      "Unauthorized": {
        "description": "Authentication required or token invalid",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Unauthorized"
                }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Not found"
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
...
      "Monitor": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "type": {
            "type": "string",
            "enum": [
              "http",
              "ping",
              "tcp",
              "keyword"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "up",
              "down",
              "degraded",
              "pending"
            ]
          },
          "interval": {
            "type": "integer",
            "description": "Check interval in seconds"
          },
          "timeout": {
            "type": "integer"
          },
          "method": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "lastCheckAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "responseTime": {
            "type": "integer",
            "description": "Last response time in ms",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "example": {
          "id": "b2c3d4e5-f6a7-8901-bcde-f12345678901",
          "name": "API Health Check",
          "url": "https://api.example.com/health",
          "type": "http",
          "status": "up",
          "interval": 300,
          "timeout": 30,
          "method": "GET",
          "active": true,
          "lastCheckAt": "2025-01-15T10:30:00Z",
          "responseTime": 142,
          "createdAt": "2025-01-01T10:00:00Z"
        }
      },
...
    }
  }
}